removed and refactored ml files
