+++ src/ml/ClassChangeModel.py
@@ -3,8 +3,6 @@ from sklearn.preprocessing import StandardScaler
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.metrics import classification_report, confusion_matrix
 import pandas as pd
-import numpy as np
-import pickle
 import joblib
 from pathlib import Path
