+++ src/ml/next_release_test.py
@@ -1,25 +1,25 @@
 import sys
 from pathlib import Path
 
-# Add the project root to Python path
 project_root = Path(__file__).parent
 sys.path.append(str(project_root))
 
-from src.ml.ClassChangeModel import ClassChangeModel
-from src.utils.ChangeTracker import ChangeTracker
+from ClassChangeModel import ClassChangeModel
+from utils.ChangeTracker import ChangeTracker
 import pandas as pd
 from sklearn.metrics import classification_report
 
-def simple_prediction_test():
+def next_release_test():
+    # hardcoded parameters for now
     end_file = 36
     mid_file = 35
     start_file = 34
-    print("=== PREDICTION TEST ===")
+    print("- PREDICTION TEST -")
     
     model = ClassChangeModel()
     model.load_model("model_trained_on_1_datasets.pkl")
     
-    path = Path('data/jgraph-jmeter-jstock-jung-lucene-weka/lucene_deps/lucene')
+    path = Path('data/jgraph-jmeter-jstock-jung-lucene-weka/lucene_deps/lucene') # hardcoded again
     tsv_files = sorted(path.glob('*.tsv'))[:end_file]  
     
     if len(tsv_files) < 2:
@@ -28,41 +28,38 @@ def simple_prediction_test():
     
     print(f"Found {len(tsv_files)} files")
     
-    # Use first 18 files for context, predict on 19th, compare with 20th
     context_files = tsv_files[:start_file]
     predict_file = tsv_files[start_file]
     
     print(f"Context files: {len(context_files)} files from {context_files[0].name} to {context_files[-1].name}")
     print(f"Predicting on: {predict_file.name}")
     
-    # Build tracker with context + prediction file
     tracker = ChangeTracker()
     for f in context_files + [predict_file]:
         df = pd.read_csv(f, sep='\t', skiprows=26)
         tracker.add_release(f.stem, df)
         print(f"  Loaded: {f.name}")
     
-    # Get data from the prediction release
     predict_release = predict_file.stem
     prediction_data = tracker.releases[predict_release]
     
     print(f"\nMaking predictions for {len(prediction_data)} classes")
     
-    # Make predictions
+    #predictions
     predictions = model.predict_changes(prediction_data)
     
-    # Now add the 20th file to get actual changes
+    # acutal changes
     if len(tsv_files) > mid_file:
         final_file = tsv_files[mid_file]
         df_final = pd.read_csv(final_file, sep='\t', skiprows=26)
         tracker.add_release(final_file.stem, df_final)
         
-        # Get actual changes that happened FROM predict_release TO final_release
-        actual_data = tracker.releases[predict_release]  # This now has "Incoming Change"
+       
+        actual_data = tracker.releases[predict_release] 
         
         print(f"Comparing with actual changes to: {final_file.name}")
         
-        # Compare predictions
+        # comparing
         correct = 0
         total = 0
         results = []
@@ -91,10 +88,9 @@ def simple_prediction_test():
         
         if total > 0:
             accuracy = correct / total
-            print(f"\n=== RESULTS ===")
+            print(f"\n- Results -")
             print(f"Accuracy: {accuracy:.3f} ({correct}/{total})")
             
-            # Show some examples
             results_df = pd.DataFrame(results)
             
             print(f"\nClassification Report:")
@@ -102,10 +98,8 @@ def simple_prediction_test():
             y_pred = results_df['Predicted'].tolist()
             print(classification_report(y_true, y_pred))
             
-            # Show correctly predicted large changes
             print(f"\nCorrectly Predicted Large Changes:")
             
-            # Filter for correct predictions of Large Decrease or Large Increase
             large_changes = results_df[
                 (results_df['Correct'] == True) & 
                 (results_df['Predicted'].isin(['Large Decrease', 'Large Increase']))
@@ -120,14 +114,12 @@ def simple_prediction_test():
                 print("--- NO LARGE CHANGES CORRECTLY PREDICTED ---")
                 print("The model did not correctly predict any Large Decrease or Large Increase cases.")
                 
-                # Show what large changes actually occurred
                 actual_large = results_df[results_df['Actual'].isin(['Large Decrease', 'Large Increase'])]
                 if len(actual_large) > 0:
                     print(f"\nActual Large Changes (Total: {len(actual_large)}):")
                     for i, (_, result) in enumerate(actual_large.head(10).iterrows()):
                         print(f"{result['Class'][:60]:<60} | Actual: {result['Actual']:<15} | Predicted: {result['Predicted']:<15}")
             
-            # Show any correctly predicted smaller changes for context
             small_changes = results_df[
                 (results_df['Correct'] == True) & 
                 (results_df['Predicted'].isin(['Small Decrease', 'Small Increase']))
@@ -139,14 +131,12 @@ def simple_prediction_test():
                 for i, (_, result) in enumerate(small_sorted.head(5).iterrows()):
                     print(f"{result['Class'][:60]:<60} | {result['Predicted']:<15} | Conf: {result['Confidence']:.3f}")
             
-            # Show accuracy by category
             print(f"\nAccuracy by Predicted Category:")
             for category in sorted(results_df['Predicted'].unique()):
                 cat_results = results_df[results_df['Predicted'] == category]
                 cat_accuracy = cat_results['Correct'].mean()
                 print(f"  {category}: {cat_accuracy:.3f} ({cat_results['Correct'].sum()}/{len(cat_results)})")
             
-            # Show actual category distribution
             print(f"\nActual Category Distribution:")
             actual_dist = results_df['Actual'].value_counts().sort_index()
             for category, count in actual_dist.items():
@@ -159,4 +149,4 @@ def simple_prediction_test():
         print("Not enough files for testing!")
 
 if __name__ == "__main__":
-    simple_prediction_test()
+    next_release_test()
